+++ fiddles/bash/fiddle-0005-ParseHtmlForJsLibs/script.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+clear;
+echo "$0" | sed 's/\.\///g' | awk '{print toupper($0)}'
+
+