+++ fiddles/three/fiddle-0002-WebGlDetection/tests/test.js
@@ -1,0 +1,11 @@
+
+TestCase("fiddle-0002-WebGlDetection", {
+    setUp: function () {
+
+    },
+    tearDown: function () {
+    },
+    "test app.util.webGl.exists": function () {
+        assertEquals(window.app.util.webGl.exists, true);
+    }
+});